% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_dbl.R
\name{age_dbl}
\alias{age_dbl}
\title{Creates a Column for Age in a numerical format}
\usage{
age_dbl(df_name, col_name)
}
\arguments{
\item{df_name}{A data frame}

\item{col_name}{A column of ages that are factors}
}
\value{
A column of ages that are numerics
}
\description{
This function creates a new column that converts factors into numbers
}
\examples{
age_dbl(snapper, Age)
}
