% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_factor_k.R
\name{cond_factor_k}
\alias{cond_factor_k}
\title{Creates a Column for the Fulton Condition Factor}
\usage{
cond_factor_k(df_name, carcass_weight, total_length)
}
\arguments{
\item{df_name}{A data frame}

\item{carcass_weight}{A column of carcass weights}

\item{total_length}{A column of total lengths}
}
\value{
A column of Fulton condition factor K
}
\description{
This function creates a new column that gives the Fulton condition
factor, K, for each observation
}
\examples{
cond_factor_k(snapper_extracted, Carcass_weight_g, Total_length_cm)
}
