% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateR.R
\name{dateR}
\alias{dateR}
\title{Creates a Column for Date in a readable format}
\usage{
dateR(df_name, col_name, date_format)
}
\arguments{
\item{df_name}{A data frame}

\item{col_name}{A column of dates}

\item{date_format}{The format the date column is in}
}
\value{
A column of dates in dttm format
}
\description{
This function creates a column in dttm format that R can more easily read
}
\examples{
dateR(snapper, DateCollected, "\%m/\%d/\%Y")
dateR(scad, DateCollected, "\%Y/\%d/\%m")
}
