% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Give Specified Statistic of Parameter}
\usage{
get_stats(df_name, parameter1, parameter2, statistic)
}
\arguments{
\item{df_name}{A data frame}

\item{parameter1}{A numeric column that you want the statistics of}

\item{parameter2}{A factor column that you want the statistics by}

\item{statistic}{The statistic you want}
}
\value{
The statistic of parameter1 based on parameter2
}
\description{
This function gives specified statistic based on the parameters supplied
}
\examples{
get_stats(snapper_extracted, 'Fork_length_cm', 'Sex', mean)
gives the mean of column 'Fork_length_cm' per 'Sex' category
get_stats(scad_extracted, 'Weight_liver_g', 'Age', sd)
gives the standard deviation of column 'Weight_liver_g' per 'Age' category
}
