% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSI.R
\name{HSI}
\alias{HSI}
\title{Creates a Column for the Hepatosomatic Index}
\usage{
HSI(df_name, liver_weight, carcass_weight)
}
\arguments{
\item{df_name}{A data frame}

\item{liver_weight}{A column of liver weights}

\item{carcass_weight}{A column of carcass weights}
}
\value{
A column of HSI
}
\description{
This function creates a new column that gives the HSI for each observation
}
\examples{
HSI(scad_full, Weight_liver_g, Carcass_weight_g)
}
