% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_col.R
\name{keep_col}
\alias{keep_col}
\title{Keep Specific Columns}
\usage{
keep_col(df_name, col_name)
}
\arguments{
\item{df_name}{A data frame}

\item{col_name}{A column name}
}
\value{
A data frame with only the specified columns
}
\description{
This function keeps specified columns.
}
\examples{
keep_col(snapper, c("MeasureValue")
keep_col(scad, c("MeasureValue", "MeasurementUnit")
}
