% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_summary.R
\name{linear_model_summary}
\alias{linear_model_summary}
\title{Summarize a linear model}
\usage{
linear_model_summary(df_name, response_variable, explanatory_variable)
}
\arguments{
\item{df_name}{A data frame}

\item{response_variable}{A column name (numeric) that is the response variable (y-axis)}

\item{explanatory_variable}{A column name (numeric) that is the explanatory variable (x-axis)}
}
\value{
A summary of the linear model. Pr(>|t|) describes the significance
}
\description{
This function provides a summary of a linear model based on the specified
variables.
}
\examples{
linear_model_summary(snapper_full, 'Total_length_cm', 'Sex_dbl')
}
