% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_col.R
\name{remove_col}
\alias{remove_col}
\title{Remove Specific Columns}
\usage{
remove_col(df_name, col_name)
}
\arguments{
\item{df_name}{A data frame}

\item{col_name}{A column name}
}
\value{
A data frame without the specified columns
}
\description{
This function removes specified columns.
}
\examples{
remove_col(snapper, c("MeasureValue")
remove_col(scad, c("MeasureValue", "MeasurementUnit")
}
