% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scad.R
\docType{data}
\name{scad}
\alias{scad}
\title{The scad animals file}
\format{
A tibble with 5240 rows and 9 columns
\describe{
\item{CaseAcc}{Factor values. Sample ID}
\item{Animal}{Factor values. Representing fish ID and common name}
\item{DateCollected}{Factor values. Representing date collected}
\item{Age}{Factor values. Representing age class}
\item{Sex}{Factor values. Representing sex class}
\item{waypoint}{Factor values. Representing location collected}
\item{Measurement}{Factor values. Representing the type of measuremnt taken}
\item{MeasureValue}{Numeric Value. Represnting the value of the measuremnt taken}
\item{MeasurementUnit}{Factor values. Representing the unit of the measurement taken}
}
}
\source{
Data collected over around Oahu
}
\usage{
scad
}
\description{
A dataset containing scad information
}
\keyword{datasets}
