% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_boxplot.R
\name{show_boxplot}
\alias{show_boxplot}
\title{Shows a boxplot}
\usage{
show_boxplot(df_name, col_range)
}
\arguments{
\item{df_name}{A data frame}

\item{col_range}{A column range. If left blank it will show all columns,
writing a single number will show a single column, writing #a:#b will show
boxplots from column#a to column#b}
}
\value{
Boxplots of the specificed columns
}
\description{
Shows a boxplot for a given column range
}
\examples{
show_boxplot(snapper,)
show_boxplot(scad, 7)
show_boxplot(scad, 7:9)
}
