% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapper_full.R
\docType{data}
\name{snapper_full}
\alias{snapper_full}
\title{The snapper extracted with numerics animals file}
\format{
A tibble with 5255 rows and 9 columns
\describe{
\item{CaseAcc}{Factor values. Sample ID}
\item{Animal}{Factor values. Representing fish ID and common name}
\item{DateCollected}{Factor values. Representing date collected}
\item{Age}{Factor values. Representing age class}
\item{Sex}{Factor values. Representing sex class}
\item{Waypoint}{Factor values. Representing location collected}
\item{Fork_length_cm}{Numeric values. Representing fork length measurement}
\item{Total_length_cm}{Numeric values. Representing toal length measurement}
\item{Weight_carcass_g}{Numeric values. Representing carcass weight measurement}
\item{Weight_liver_g}{Numeric values. Representing liver weight measurement}
\item{Waypoint_dbl}{Numeric Values. Representing waypoints as numbers}
\item{Age_dbl}{Numeric Values. Representing ages as numbers}
\item{Sex_dbl}{Numeric Values. Representing sexes as numbers}
\item{DateR}{Dttm Values. Representing date collected as dttm}
}
}
\source{
Data collected over around Oahu
}
\usage{
snapper_full
}
\description{
A dataset containing snapper information with repeated observations extracted
and grouped together and factor columns converted into numerics
}
\keyword{datasets}
