% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_col_value.R
\name{subset_by_col_value}
\alias{subset_by_col_value}
\title{Extract Observations by Column Value}
\usage{
subset_by_col_value(df_name, col_name, col_val)
}
\arguments{
\item{df_name}{A data frame}

\item{col_name}{A column name}

\item{col_val}{The item being filtered for, in the form of 'col_value'}
}
\value{
A data frame with only the specified col_value
}
\description{
This function takes only the observations in a table that match
the specified column value in the specified column.
}
\examples{
subset_by_col_value(snapper, MeasureValue, 'Fork length')
subset_by_col_value(scad, MeasureValue, 'Weight liver')
}
