% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_explan_mlm_significance.R
\name{three_explan_mlm_significance}
\alias{three_explan_mlm_significance}
\title{Show significance of a linear model with three explanatory variables}
\usage{
three_explan_mlm_significance(
  df_name,
  response_variable,
  explanatory_variable1,
  explanatory_variable2,
  explanatory_variable3
)
}
\arguments{
\item{df_name}{A data frame}

\item{response_variable}{A column name (numeric) that is the response variable (y-axis)}

\item{explanatory_variable1}{A column name (numeric) that is the first explanatory variable (x-axis)}

\item{explanatory_variable2}{A column name (numeric) that is the second explanatory variable (x-axis)}

\item{explanatory_variable3}{A column name (numeric) that is the third explanatory variable (x-axis)}
}
\value{
A stepAIC of the explanatory variables. The variables displayed
with the smallest AIC are significant.
}
\description{
This function provides the cosignificance of a linear model based
on the specified variables.
}
\examples{
three_explan_mlm_significance(snapper_full, 'Total_length_cm', 'Sex_dbl', 'Age_dbl', 'Waypoint_dbl')
}
