% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waypoint_dbl.R
\name{waypoint_dbl}
\alias{waypoint_dbl}
\title{Creates a Column for Waypoints in a numerical format}
\usage{
waypoint_dbl(df_name, col_name)
}
\arguments{
\item{df_name}{A data frame}

\item{col_name}{A column of waypoints that are factors}
}
\value{
A column of waypoints that are numerics
}
\description{
This function creates a new column that converts factors into numbers
}
\examples{
waypoint_dbl(snapper, waypoint)
}
